chore: wip

+++ examples/move_windows_then_back.rs
@@ -285,19 +285,19 @@ fn main() {
 
         // Create animation
         let animation = WindowAnimation::new(
-            *hwnd,
-            *original_rect,
-            grid_rect,
-            Duration::from_millis(duration_ms),
-            easing_type,
-        );
+                    *hwnd as u64,
+                    *original_rect,
+                    grid_rect,
+                    Duration::from_millis(duration_ms),
+                    easing_type,
+                );
 
         animations.push(animation);
 
         // Start the animation by moving window to start position and then animate
         tracker
             .start_window_animation(
-                *hwnd,
+                *hwnd as u64,
                 grid_rect,
                 Duration::from_millis(duration_ms),
                 easing_type,
@@ -366,12 +366,12 @@ fn main() {
         }
 
         let return_animation = WindowAnimation::new(
-            *hwnd,
-            current_rect,
-            *original_rect,
-            Duration::from_millis(duration_ms),
-            easing_type,
-        );
+                    *hwnd as u64,
+                    current_rect,
+                    *original_rect,
+                    Duration::from_millis(duration_ms),
+                    easing_type,
+                );
 
         // Animate back to original position
         let animation_start = std::time::Instant::now();
